Remove settings folder
