--- settings/server.js
@@ -1,3 +1,0 @@
-module.exports = {
-  port: parseInt(process.env.PORT, 10) || 8080,
-};
